Create Task-3A.ino
Utilize all the keys on the keypad and add functions. The functionality should be
remap "*" and "#" to be 'E' and 'F'.
Enter Hexadecimal values on the keypad.
press a push button attached seperately to trigger that the Hex phrase has been entered and is ready to be processed and displayed.